+++ apps/restconf/restconf_methods.h
@@ -45,8 +45,13 @@ int api_data_options(clicon_handle h, FCGX_Request *r);
 int api_data_put(clicon_handle h, FCGX_Request *r, char *api_path, 
 		 cvec *pcvec, int pi, 
 		 cvec *qvec, char *data,
-		 int pretty,
-		 restconf_media media_in,     restconf_media media_out);
+		 int pretty, restconf_media media_out);
+
+int api_data_patch(clicon_handle h, FCGX_Request *r, char *api_path, 
+		   cvec *pcvec, int pi, 
+		   cvec *qvec, char *data, int pretty,
+		   restconf_media media_out);
+
 int api_data_delete(clicon_handle h, FCGX_Request *r, char *api_path, int pi,
 		    int pretty, restconf_media media_out);