+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Major New features
 * Restconf RFC 8040 increased feature compliance
+  * RESTCONF PATCH (plain patch) is supported according to RFC 8040 4.6.1
   * RESTCONF "insert" and "point" query parameters supported
     * Applies to ordered-by-user leaf and leaf-lists
   * RESTCONF PUT/POST erroneously returned 200 OK. Instead restconf now returns: